refactor: consolidate path-basics helpers into @utils/core
normalizeFilePath/getBasename/getFileStem lived in src/shared/utils/path.ts,
a directory documented as "schemas and message handlers across webviews."
In practice these are generic, dependency-free string helpers used by 20+
backend, CLI, desktop, and webview files, giving the codebase two homes for
browser-safe generic utilities (src/shared/utils/ and src/utils/core/) with
no clear rule for which owns path helpers.

Move the three functions into src/utils/core/index.ts (the existing,
self-documented home for framework-free helpers shared by webview and
backend code) alongside pathCore.ts's Node-dependent path utilities, delete
src/shared/utils/path.ts, and repoint all call sites at @utils/core. Merge
the getBasename test suite into UtilsCore.vitest.ts.

No behavior change — verified via full typecheck, all three webview Vite
bundles, and the vitest suite (pre-existing, unrelated process-signal test
failure aside).
