+++ src/agent/utils/outputFileUtils.ts
@@ -4,7 +4,7 @@ import {
   WORKFLOW_OUTPUT_BASENAME,
   workflowOutputPath,
 } from '@agent/output/workflowOutputLayout';
-import { normalizeFilePath } from '@shared/utils/path';
+import { normalizeFilePath } from '@utils/core';
 
 /**
  * Generates a runDir-relative output path under a round subfolder:
