+++ src/test-kernel/utils/UtilsCore.vitest.ts
@@ -1,5 +1,8 @@
 // Suites for @utils/core (comparators, type guards, async helpers).
 
+// Standard library imports
+import * as assert from 'node:assert';
+
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import {
   createFlushableDebounce,
@@ -7,6 +10,7 @@ import {
   ensureArray,
   filterNotNull,
   filterNotNullish,
+  getBasename,
   toNewestFirstByTimestamp,
 } from '@utils/core';
 
@@ -72,6 +76,49 @@ describe('core type guard predicates', () => {
 
   it('wraps scalar values in an array', () => {
     expect(ensureArray('alpha')).toEqual(['alpha']);
+  });
+});
+
+// ---------------------------------------------------------------------------
+// pathBasics
+// ---------------------------------------------------------------------------
+
+describe('getBasename', () => {
+  it.each([
+    ['/home/user/file.txt', 'file.txt'],
+    ['/usr/local/bin/node', 'node'],
+    ['/path/to/document.pdf', 'document.pdf'],
+    ['C:\\Users\\file.txt', 'file.txt'],
+    ['C:\\Program Files\\app.exe', 'app.exe'],
+    ['D:\\Documents\\report.docx', 'report.docx'],
+    ['C:/Users\\Documents/file.txt', 'file.txt'],
+    ['/home\\user/document.pdf', 'document.pdf'],
+    ['/path/to/', 'to'],
+    ['/path/to/dir/', 'dir'],
+    ['C:\\Users\\', 'Users'],
+    ['', ''],
+    ['/', ''],
+    ['//', ''],
+    ['file.txt', 'file.txt'],
+    ['./file.txt', 'file.txt'],
+    ['../file.txt', 'file.txt'],
+    ['/path/to/file.tar.gz', 'file.tar.gz'],
+    ['archive.backup.zip', 'archive.backup.zip'],
+    ['/home/user/.bashrc', '.bashrc'],
+    ['/path/to/file with spaces.txt', 'file with spaces.txt'],
+    ['/path/to/file-with-dashes.txt', 'file-with-dashes.txt'],
+    ['/path/to/file_with_underscores.txt', 'file_with_underscores.txt'],
+    ['relative/path/to/file.txt', 'file.txt'],
+    ['./relative/file.txt', 'file.txt'],
+    ['../parent/file.txt', 'file.txt'],
+    ['/home/user/Documents', 'Documents'],
+    ['C:\\Program Files', 'Program Files'],
+    ['/usr/local/bin', 'bin'],
+    // Regression: paths ending with a separator used to return empty.
+    ['folder/', 'folder'],
+    ['/home/user/folder/', 'folder'],
+  ])('getBasename(%j) === %j', (input, expected) => {
+    assert.strictEqual(getBasename(input), expected);
   });
 });
