+++ src/tools/executions/runDirectoryFiles.ts
@@ -6,7 +6,7 @@
 
 import * as path from 'node:path';
 
-import { normalizeFilePath } from '@shared/utils/path';
+import { normalizeFilePath } from '@utils/core';
 import { StorageFS } from '@utils/files';
 import { isDirectory } from '@utils/files/fsEntryType';
